Configure linter with airbnb config. Added TestScene. Moved shaders and libs to corresponding folder for linter ignore. Updated react-scrits.
