+++ .gitignore
@@ -8,6 +8,7 @@
 /build
 
 # misc
+.env
 .DS_Store
 .env.local
 .env.development.local
