+++ src/index.js
@@ -1,5 +1,17 @@
 import initWebScene from './HKAlley';
+import initTestScene from './TestScene';
+
+const scene = process.env.REACT_APP_SCENE;
 
 document.body.style.margin = '0';
 
-initWebScene();
+switch (scene) {
+  case 'main':
+    initWebScene();
+    break;
+  case 'test':
+    initTestScene();
+    break;
+  default:
+    initWebScene();
+}
