+++ src/libs/rectAreaLights.js
